feat: Add opening balance field and update stock quantity type to INT in product-related entities and forms

+++ backend/src/main/java/com/reallink/pump/services/ProductService.java
@@ -139,8 +139,8 @@ public class ProductService {
                     existingProduct,
                     oldSalesUnit,
                     newSalesUnit,
-                    existingProduct.getStockQuantity(), // old stock quantity
-                    existingProduct.getStockQuantity(), // new stock quantity unchanged here
+                    BigDecimal.valueOf(existingProduct.getStockQuantity()), // old stock quantity
+                    BigDecimal.valueOf(existingProduct.getStockQuantity()), // new stock quantity unchanged here
                     oldSalesRate,
                     newSalesRate,
                     reason,
